fix: OAuth2 설정 변경

+++ closet/src/main/java/com/codeit/closet/common/security/jwt/JwtTokenProvider.java
@@ -100,7 +100,7 @@ public class JwtTokenProvider {
         .secure(false)
         .sameSite("None")
         .path("/")
-        .maxAge(refreshTokenExpirationMs / 1000)
+        .maxAge((int) (refreshTokenExpirationMs / 1000L))
         .build();
   }
